requirements.txt
# ETL Agents Framework Requirements

# Data Processing
pandas>=1.3.0
numpy>=1.20.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
missingno>=0.5.0

# Data Science
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Web Framework
Flask>=2.0.0
Flask-CORS>=3.0.0
Werkzeug>=2.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.19.0
tqdm>=4.62.0

# Logging
loguru>=0.5.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
